Now user can't edit/delete his "master" category - the one named after his username.

+++ src/willy/gallery/views.py
@@ -34,6 +34,10 @@ def edit_category(request, category_id):
         return render_to_response('gallery_error.html',
                                   {'message' : _("You don't have enough privileges to edit this category")},
                                   context_instance=RequestContext(request))
+    if category.name == category.owner.username:
+        return render_to_response('gallery_error.html',
+                                  {'message' : _("You can't delete this category")},
+                                  context_instance=RequestContext(request))
     if request.method == 'GET':
         return render_to_response('edit_category.html',
                                   {'form' : CategoryForm(instance=category),
@@ -60,6 +64,10 @@ def delete_category(request, category_id):
     if category.owner != request.user:
         return render_to_response('gallery_error.html',
                                   {'message' : _("You don't have enough privileges to delete this category")},
+                                  context_instance=RequestContext(request))
+    if category.name == category.owner.username:
+        return render_to_response('gallery_error.html',
+                                  {'message' : _("You can't delete this category")},
                                   context_instance=RequestContext(request))
 
     category.delete()
